+++ examples/src/bin/futures.rs
@@ -1,6 +1,7 @@
-use async_imap::error::{Error, Result};
-use async_std::prelude::*;
 use std::env;
+
+use async_imap::error::{Error, Result};
+use futures::TryStreamExt;
 
 fn main() -> Result<()> {
     futures::executor::block_on(async {
@@ -36,7 +37,7 @@ async fn fetch_inbox_top(imap_server: &str, login: &str, password: &str) -> Resu
     // fetch message number 1 in this mailbox, along with its RFC822 field.
     // RFC 822 dictates the format of the body of e-mails
     let messages_stream = imap_session.fetch("1", "RFC822").await?;
-    let messages: Vec<_> = messages_stream.collect::<Result<_>>().await?;
+    let messages: Vec<_> = messages_stream.try_collect().await?;
     let message = if let Some(m) = messages.first() {
         m
     } else {
