+++ src/client.rs
@@ -4,13 +4,21 @@ use std::ops::{Deref, DerefMut};
 use std::pin::Pin;
 use std::str;
 
+use async_channel::{self as channel, bounded};
 use async_native_tls::{TlsConnector, TlsStream};
-use async_std::channel;
-use async_std::net::{TcpStream, ToSocketAddrs};
+#[cfg(feature = "runtime-async-std")]
+use async_std::{
+    io::{Read, Write, WriteExt},
+    net::{TcpStream, ToSocketAddrs},
+};
 use extensions::quota::parse_get_quota_root;
-use futures::io::{self, AsyncRead as Read, AsyncWrite as Write};
-use futures::prelude::*;
+use futures::{io, Stream, StreamExt};
 use imap_proto::{RequestId, Response};
+#[cfg(feature = "runtime-tokio")]
+use tokio::{
+    io::{AsyncRead as Read, AsyncWrite as Write, AsyncWriteExt},
+    net::{TcpStream, ToSocketAddrs},
+};
 
 use super::authenticator::Authenticator;
 use super::error::{Error, ParseError, Result, ValidateError};
@@ -359,7 +367,7 @@ impl<T: Read + Write + Unpin + fmt::Debug + Send> Session<T> {
 
     // not public, just to avoid duplicating the channel creation code
     fn new(conn: Connection<T>) -> Self {
-        let (tx, rx) = channel::bounded(100);
+        let (tx, rx) = bounded(100);
         Session {
             conn,
             unsolicited_responses: rx,
@@ -814,12 +822,15 @@ impl<T: Read + Write + Unpin + fmt::Debug + Send> Session<T> {
     ///
     /// ```no_run
     /// use async_imap::{types::Seq, Session, error::Result};
-    /// use async_std::prelude::*;
+    /// #[cfg(feature = "runtime-async-std")]
     /// use async_std::net::TcpStream;
+    /// #[cfg(feature = "runtime-tokio")]
+    /// use tokio::net::TcpStream;
+    /// use futures::TryStreamExt;
     ///
     /// async fn delete(seq: Seq, s: &mut Session<TcpStream>) -> Result<()> {
     ///     let updates_stream = s.store(format!("{}", seq), "+FLAGS (\\Deleted)").await?;
-    ///     let _updates: Vec<_> = updates_stream.collect::<Result<_>>().await?;
+    ///     let _updates: Vec<_> = updates_stream.try_collect().await?;
     ///     s.expunge().await?;
     ///     Ok(())
     /// }
@@ -1464,6 +1475,7 @@ mod tests {
     use super::super::mock_stream::MockStream;
     use super::*;
     use std::borrow::Cow;
+    use std::future::Future;
 
     use async_std::sync::{Arc, Mutex};
     use imap_proto::Status;
@@ -1490,7 +1502,8 @@ mod tests {
         };
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn fetch_body() {
         let response = "a0 OK Logged in.\r\n\
                         * 2 FETCH (BODY[TEXT] {3}\r\nfoo)\r\n\
@@ -1500,7 +1513,8 @@ mod tests {
         session.read_response().await.unwrap().unwrap();
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn readline_delay_read() {
         let greeting = "* OK Dovecot ready.\r\n";
         let mock_stream = MockStream::default()
@@ -1519,7 +1533,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn readline_eof() {
         let mock_stream = MockStream::default().with_eof();
         let mut client = mock_client!(mock_stream);
@@ -1527,7 +1542,8 @@ mod tests {
         assert!(res.is_none());
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     #[should_panic]
     async fn readline_err() {
         // TODO Check the error test
@@ -1536,7 +1552,8 @@ mod tests {
         client.read_response().await.unwrap().unwrap();
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn authenticate() {
         let response = b"+ YmFy\r\n\
                          A0001 OK Logged in\r\n"
@@ -1567,7 +1584,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn login() {
         let response = b"A0001 OK Logged in\r\n".to_vec();
         let username = "username";
@@ -1586,7 +1604,8 @@ mod tests {
         }
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn logout() {
         let response = b"A0001 OK Logout completed.\r\n".to_vec();
         let command = "A0001 LOGOUT\r\n";
@@ -1599,7 +1618,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn rename() {
         let response = b"A0001 OK RENAME completed\r\n".to_vec();
         let current_mailbox_name = "INBOX";
@@ -1621,7 +1641,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn subscribe() {
         let response = b"A0001 OK SUBSCRIBE completed\r\n".to_vec();
         let mailbox = "INBOX";
@@ -1635,7 +1656,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn unsubscribe() {
         let response = b"A0001 OK UNSUBSCRIBE completed\r\n".to_vec();
         let mailbox = "INBOX";
@@ -1649,7 +1671,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn expunge() {
         let response = b"A0001 OK EXPUNGE completed\r\n".to_vec();
         let mock_stream = MockStream::new(response);
@@ -1661,7 +1684,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_expunge() {
         let response = b"* 2 EXPUNGE\r\n\
             * 3 EXPUNGE\r\n\
@@ -1682,7 +1706,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn check() {
         let response = b"A0001 OK CHECK completed\r\n".to_vec();
         let mock_stream = MockStream::new(response);
@@ -1694,7 +1719,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn examine() {
         let response = b"* FLAGS (\\Answered \\Flagged \\Deleted \\Seen \\Draft)\r\n\
             * OK [PERMANENTFLAGS ()] Read-only mailbox.\r\n\
@@ -1733,7 +1759,8 @@ mod tests {
         assert_eq!(mailbox, expected_mailbox);
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn select() {
         let response = b"* FLAGS (\\Answered \\Flagged \\Deleted \\Seen \\Draft)\r\n\
             * OK [PERMANENTFLAGS (\\* \\Answered \\Flagged \\Deleted \\Draft \\Seen)] \
@@ -1781,7 +1808,8 @@ mod tests {
         assert_eq!(mailbox, expected_mailbox);
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn search() {
         let response = b"* SEARCH 1 2 3 4 5\r\n\
             A0001 OK Search completed\r\n"
@@ -1797,7 +1825,8 @@ mod tests {
         assert_eq!(ids, [1, 2, 3, 4, 5].iter().cloned().collect());
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_search() {
         let response = b"* SEARCH 1 2 3 4 5\r\n\
             A0001 OK Search completed\r\n"
@@ -1813,7 +1842,8 @@ mod tests {
         assert_eq!(ids, [1, 2, 3, 4, 5].iter().cloned().collect());
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_search_unordered() {
         let response = b"* SEARCH 1 2 3 4 5\r\n\
             A0002 OK CAPABILITY completed\r\n\
@@ -1830,7 +1860,8 @@ mod tests {
         assert_eq!(ids, [1, 2, 3, 4, 5].iter().cloned().collect());
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn capability() {
         let response = b"* CAPABILITY IMAP4rev1 STARTTLS AUTH=GSSAPI LOGINDISABLED\r\n\
             A0001 OK CAPABILITY completed\r\n"
@@ -1849,7 +1880,8 @@ mod tests {
         }
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn create() {
         let response = b"A0001 OK CREATE completed\r\n".to_vec();
         let mailbox_name = "INBOX";
@@ -1863,7 +1895,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn delete() {
         let response = b"A0001 OK DELETE completed\r\n".to_vec();
         let mailbox_name = "INBOX";
@@ -1877,7 +1910,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn noop() {
         let response = b"A0001 OK NOOP completed\r\n".to_vec();
         let mock_stream = MockStream::new(response);
@@ -1889,7 +1923,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn close() {
         let response = b"A0001 OK CLOSE completed\r\n".to_vec();
         let mock_stream = MockStream::new(response);
@@ -1901,7 +1936,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn store() {
         generic_store(" ", |c, set, query| async move {
             c.lock()
@@ -1915,7 +1951,8 @@ mod tests {
         .await;
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_store() {
         generic_store(" UID ", |c, set, query| async move {
             c.lock()
@@ -1942,7 +1979,8 @@ mod tests {
         generic_with_uid(res, "STORE", "2.4", "+FLAGS (\\Deleted)", prefix, op).await;
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn copy() {
         generic_copy(" ", |c, set, query| async move {
             c.lock().await.copy(set, query).await?;
@@ -1951,7 +1989,8 @@ mod tests {
         .await;
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_copy() {
         generic_copy(" UID ", |c, set, query| async move {
             c.lock().await.uid_copy(set, query).await?;
@@ -1976,7 +2015,8 @@ mod tests {
         .await;
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn mv() {
         let response = b"* OK [COPYUID 1511554416 142,399 41:42] Moved UIDs.\r\n\
             * 2 EXPUNGE\r\n\
@@ -1994,7 +2034,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_mv() {
         let response = b"* OK [COPYUID 1511554416 142,399 41:42] Moved UIDs.\r\n\
             * 2 EXPUNGE\r\n\
@@ -2012,7 +2053,8 @@ mod tests {
         );
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn fetch() {
         generic_fetch(" ", |c, seq, query| async move {
             c.lock()
@@ -2027,7 +2069,8 @@ mod tests {
         .await;
     }
 
-    #[async_std::test]
+    #[cfg_attr(feature = "runtime-tokio", tokio::test)]
+    #[cfg_attr(feature = "runtime-async-std", async_std::test)]
     async fn uid_fetch() {
         generic_fetch(" UID ", |c, seq, query| async move {
             c.lock()
